+++ py/nodes/misc.py
@@ -1,6 +1,9 @@
+from __future__ import annotations
+
 import random
 
 import torch
+from comfy import model_management
 
 
 class DiscardPenultimateSigma:
@@ -19,20 +22,31 @@ class DiscardPenultimateSigma:
     DESCRIPTION = "Discards the next to last sigma in the list."
 
     @classmethod
-    def go(cls, enabled, sigmas):
+    def go(cls, enabled: bool, sigmas: torch.Tensor) -> tuple[torch.Tensor]:
         if not enabled or len(sigmas) < 2:
             return (sigmas,)
         return (torch.cat((sigmas[:-2], sigmas[-1:])),)
 
 
 class SeededDisableNoise:
-    def __init__(self, seed):
+    def __init__(self, seed: int, seed_offset: int = 1):
         self.seed = seed
+        self.seed_offset = seed_offset
 
     def generate_noise(self, latent):
         samples = latent["samples"]
         torch.manual_seed(self.seed)
         random.seed(self.seed)  # For good measure.
+        shape = samples.shape
+        device = model_management.get_torch_device()
+        for _ in range(self.seed_offset):
+            _ = random.random()  # noqa: S311
+            _ = torch.randn(
+                *shape,
+                dtype=samples.dtype,
+                layout=samples.layout,
+                device=device,
+            )
         return torch.zeros(
             samples.shape,
             dtype=samples.dtype,
@@ -56,11 +70,31 @@ class BlehDisableNoise:
                     {"default": 0, "min": 0, "max": 0xFFFFFFFFFFFFFFFF},
                 ),
             },
+            "optional": {
+                "seed_offset": (
+                    "INT",
+                    {
+                        "default": 1,
+                        "min": 0,
+                        "max": 200,
+                        "tooltip": "Advances the RNG this many times to avoid the mistake of using the same noise for sampling as the initial noise. I recommend leaving this at 1 (or higher) but you can set it to 0. to disable",
+                    },
+                ),
+            },
         }
 
     @classmethod
-    def go(cls, noise_seed):
-        return (SeededDisableNoise(noise_seed),)
+    def go(
+        cls,
+        noise_seed: int,
+        seed_offset: None | int = 1,
+    ) -> tuple[SeededDisableNoise]:
+        return (
+            SeededDisableNoise(
+                noise_seed,
+                seed_offset if seed_offset is not None else 0,
+            ),
+        )
 
 
 class Wildcard(str):